requirements.txt
python-telegram-bot==13.15
pydub
numpy
matplotlib
Pillow
ffmpeg-python
